+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/RequestParameterFrom.kt
@@ -1,5 +1,5 @@
 package at.asitplus.openid
-
-interface RequestParametersFrom {
-    val parameters: RequestParameters
-}
+//
+//interface RequestParametersFrom <T:RequestParameters>{
+//    val parameters: T
+//}