+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/helper/RequestParser.kt
@@ -3,7 +3,6 @@ package at.asitplus.wallet.lib.oidc.helper
 import at.asitplus.KmmResult
 import at.asitplus.catching
 import at.asitplus.openid.AuthenticationRequestParameters
-import at.asitplus.openid.AuthenticationRequestParametersFrom
 import at.asitplus.openid.AuthenticationResponseParameters
 import at.asitplus.openid.OpenIdConstants
 import at.asitplus.openid.OpenIdConstants.Errors
@@ -11,6 +10,7 @@ import at.asitplus.openid.RequestParameters
 import at.asitplus.openid.RequestParametersFrom
 import at.asitplus.signum.indispensable.josef.JsonWebKeySet
 import at.asitplus.signum.indispensable.josef.JwsSigned
+import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.oidc.AuthenticationResponseResult
 import at.asitplus.wallet.lib.oidc.RemoteResourceRetrieverFunction
 import at.asitplus.wallet.lib.oidc.RequestObjectJwsVerifier
@@ -24,7 +24,7 @@ import io.ktor.util.*
 import kotlinx.serialization.PolymorphicSerializer
 import kotlinx.serialization.json.JsonObject
 
-open class RequestParser(
+class RequestParser(
     /**
      * Need to implement if resources are defined by reference, i.e. the URL for a [JsonWebKeySet],
      * or the request itself as `request_uri`, or `presentation_definition_uri`.
@@ -36,14 +36,14 @@ open class RequestParser(
      * Need to verify the request object serialized as a JWS,
      * which may be signed with a pre-registered key (see [OpenIdConstants.ClientIdScheme.PreRegistered]).
      */
-    private val requestObjectJwsVerifier: RequestObjectJwsVerifier = RequestObjectJwsVerifier { _, _: Any -> true },
+    private val requestObjectJwsVerifier: RequestObjectJwsVerifier = RequestObjectJwsVerifier { _: Any -> true },
 ) {
     /**
      * Pass in the URL sent by the Verifier (containing the [RequestParameters] as query parameters),
      * to create [AuthenticationResponseParameters] that can be sent back to the Verifier, see
      * [AuthenticationResponseResult].
      */
-    suspend fun parseRequestParameters(input: String): KmmResult<RequestParametersFrom> = catching {
+    suspend fun parseRequestParameters(input: String): KmmResult<RequestParametersFrom<*>> = catching {
         // maybe it is a request JWS
         val parsedParams = kotlin.run { parseRequestObjectJws(input) }
             ?: kotlin.runCatching { // maybe it's in the URL parameters
@@ -70,7 +70,7 @@ open class RequestParser(
         extractedParams
     }
 
-    private suspend fun extractRequestObject(params: AuthenticationRequestParameters): RequestParametersFrom? =
+    private suspend fun extractRequestObject(params: AuthenticationRequestParameters): RequestParametersFrom<*>? =
         params.request?.let { requestObject ->
             parseRequestObjectJws(requestObject)
         } ?: params.requestUri?.let { uri ->
@@ -78,35 +78,27 @@ open class RequestParser(
                 ?.let { parseRequestParameters(it).getOrNull() }
         }
 
-    private fun parseRequestObjectJws(requestObject: String): RequestParametersFrom? {
-        return JwsSigned.deserialize(requestObject).getOrNull()?.let { jws ->
-            val params = kotlin.runCatching {
-                jsonSerializer.decodeFromString(
-                    PolymorphicSerializer(RequestParameters::class),
-                    jws.payload.decodeToString()
-                )
-            }.getOrElse {
-                return null
-                    .apply { Napier.w("parseRequestObjectJws: Deserialization failed", it) }
+    private fun parseRequestObjectJws(requestObject: String): RequestParametersFrom<*>? {
+        return JwsSigned.deserialize<AuthenticationRequestParameters>(requestObject, vckJsonSerializer).getOrNull()
+            ?.let { jws ->
+                if (requestObjectJwsVerifier.invoke(jws)) {
+                    RequestParametersFrom.JwsSigned(jws, jws.payload)
+                } else null
+                    .also { Napier.w("parseRequestObjectJws: Signature not verified for $jws") }
             }
-            if (requestObjectJwsVerifier.invoke(jws, params)) {
-                matchRequestParameterCases(jws, params)
-            } else null
-                .also { Napier.w("parseRequestObjectJws: Signature not verified for $jws") }
-        }
     }
 
-    open fun <T> matchRequestParameterCases(input: T, params: RequestParameters): RequestParametersFrom =
+    private fun <T> matchRequestParameterCases(input: T, params: RequestParameters): RequestParametersFrom<*> =
         when (params) {
             is AuthenticationRequestParameters ->
                 when (input) {
-                    is Url -> AuthenticationRequestParametersFrom.Uri(input, params)
-                    is JwsSigned<*> -> AuthenticationRequestParametersFrom.JwsSigned(
-                        input as JwsSigned<ByteArray>,
+                    is Url -> RequestParametersFrom.Uri(input, params)
+                    is JwsSigned<*> -> RequestParametersFrom.JwsSigned(
+                        input as JwsSigned<RequestParameters>,
                         params
                     )
 
-                    is String -> AuthenticationRequestParametersFrom.Json(input, params)
+                    is String -> RequestParametersFrom.Json(input, params)
                     else -> throw Exception("matchRequestParameterCases: unknown type ${input?.let { it::class.simpleName } ?: "null"}")
                 }